fix(Minor fixes): Small fixes, composite subscriptions, moved some files to correct package
