+++ app/src/main/java/info/blockchain/wallet/viewModel/MainViewModel.java
@@ -32,6 +32,7 @@ import javax.inject.Inject;
 
 import piuk.blockchain.android.di.Injector;
 import rx.Observable;
+import rx.subscriptions.CompositeSubscription;
 
 @SuppressWarnings("WeakerAccess")
 public class MainViewModel implements ViewModel {
@@ -39,6 +40,7 @@ public class MainViewModel implements ViewModel {
     private Context context;
     private DataListener dataListener;
     private OSUtil osUtil;
+    private CompositeSubscription compositeSubscription;
     @Inject protected PrefsUtil prefs;
     @Inject protected AppUtil appUtil;
     @Inject protected PayloadManager payloadManager;
@@ -64,6 +66,7 @@ public class MainViewModel implements ViewModel {
         this.dataListener = dataListener;
         this.osUtil = new OSUtil(context);
         this.appUtil.applyPRNGFixes();
+        this.compositeSubscription = new CompositeSubscription();
     }
 
     public void onViewReady() {
@@ -75,15 +78,16 @@ public class MainViewModel implements ViewModel {
 
     private void checkIfShouldShowEmailVerification() {
         if (appUtil.isNewlyCreated()) {
-            getSettingsApi()
-                    .compose(RxUtil.applySchedulers())
-                    .subscribe(settings -> {
-                        if (!settings.isEmailVerified()) {
-                            appUtil.setNewlyCreated(false);
-                            String email = settings.getEmail();
-                            dataListener.showEmailVerificationDialog(email);
-                        }
-                    }, Throwable::printStackTrace);
+            compositeSubscription.add(
+                    getSettingsApi()
+                            .compose(RxUtil.applySchedulers())
+                            .subscribe(settings -> {
+                                if (!settings.isEmailVerified()) {
+                                    appUtil.setNewlyCreated(false);
+                                    String email = settings.getEmail();
+                                    dataListener.showEmailVerificationDialog(email);
+                                }
+                            }, Throwable::printStackTrace));
         }
     }
 
@@ -182,6 +186,7 @@ public class MainViewModel implements ViewModel {
         dataListener = null;
         stopWebSocketService();
         DynamicFeeCache.getInstance().destroy();
+        compositeSubscription.clear();
     }
 
     private void exchangeRateThread() {
